+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/AbstractSummaryAction.cs
@@ -24,7 +24,7 @@ namespace Azure.AI.TextAnalytics
         {
             ModelVersion = options.ModelVersion;
             DisableServiceLogs = options.DisableServiceLogs;
-            MaxSentenceCount = options.MaxSentenceCount;
+            SentenceCount = options.SentenceCount;
         }
 
         /// <summary>
@@ -52,9 +52,9 @@ namespace Azure.AI.TextAnalytics
         public string ActionName { get; set; }
 
         /// <summary>
-        /// The maximum number of sentences that the resulting summaries can have. If not set, the service default is
-        /// used.
+        /// The desired number of sentences in the resulting summaries, which the service will attempt to approximate.
+        /// If not set, the service default is used.
         /// </summary>
-        public int? MaxSentenceCount { get; set; }
+        public int? SentenceCount { get; set; }
     }
 }